+++ src/LASlib/lasreader_ply.cpp
@@ -562,7 +562,7 @@ void LASreaderPLY::add_attribute(I32 attribute_type, const char* name, const cha
   else
   {
     char temp[32];
-    sprintf(temp, "attribute %d", number_attributes);
+    snprintf(temp, 32, "attribute %d", number_attributes);
     attribute_names[number_attributes] = LASCopyString(temp);
   }
   if (description)