+++ NEWS.md
@@ -1,6 +1,7 @@
 ### rlas v1.6.2
 
 - Fix compilation with g++-12 thanks to Dirk Eddelbuettel [#57](https://github.com/r-lidar/rlas/pull/57)
+- Replace `sprintf` with `snprintf`
 
 ### rlas v1.6.1